+++ backend/internal/db/mongo/carousel_text_crud.go
@@ -184,7 +184,7 @@ func (b *Backend) GetDeletedCarouselTexts(page uint64, size uint64) ([]*models.C
 	return accs, nil
 }
 
-func (b *Backend) CountDeletedCarouselTexts() (int64, error) {
+func (b *Backend) CountDeletedCarouselTexts() (uint64, error) {
 	ctx, cancel := b.GetContext()
 	defer cancel()
 
@@ -197,5 +197,5 @@ func (b *Backend) CountDeletedCarouselTexts() (int64, error) {
 		return 0, err
 	}
 
-	return count, nil
+	return uint64(count), nil
 }
