+++ backend/internal/db/mongo/refill_crud.go
@@ -184,7 +184,7 @@ func (b *Backend) GetDeletedRefills(page uint64, size uint64) ([]*models.Refill,
 	return accs, nil
 }
 
-func (b *Backend) CountDeletedRefills() (int64, error) {
+func (b *Backend) CountDeletedRefills() (uint64, error) {
 	ctx, cancel := b.GetContext()
 	defer cancel()
 
@@ -197,5 +197,5 @@ func (b *Backend) CountDeletedRefills() (int64, error) {
 		return 0, err
 	}
 
-	return count, nil
+	return uint64(count), nil
 }
